Refactoring, added Adapter, added icons, started testing

+++ app/src/main/java/truebeans/fyruz/meteofy/Utils/JsonParsing/Main.kt
@@ -13,11 +13,10 @@ For support, please feel free to contact me at https://www.linkedin.com/in/syeda
 
 
 data class Main (
-
 	@SerializedName("temp") val temp : Double,
 	@SerializedName("feels_like") val feels_like : Double,
 	@SerializedName("temp_min") val temp_min : Double,
-	@SerializedName("temp_max") val temp_max : Int,
-	@SerializedName("pressure") val pressure : Int,
-	@SerializedName("humidity") val humidity : Int
+	@SerializedName("temp_max") val temp_max : Double,
+	@SerializedName("pressure") val pressure : Double,
+	@SerializedName("humidity") val humidity : Double
 )
